StarMapEntryVanilla distance can be byte

+++ oni-seed-browser-model/src/commonMain/kotlin/de/stefan_oltmann/oni/model/StarMapEntryVanilla.kt
@@ -30,5 +30,5 @@ data class StarMapEntryVanilla(
     val id: VanillaSpacePOI,
 
     @ProtoNumber(2)
-    val distance: Int
+    val distance: Byte
 )
